+++ ricaun.RevitTest.TestAdapter/TestAdapter.cs
@@ -5,6 +5,7 @@
 
 using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
+using ricaun.RevitTest.TestAdapter.Models;
 using System;
 using System.Reflection;
 
@@ -31,7 +32,34 @@ namespace ricaun.RevitTest.TestAdapter
 
             AdapterLogger.Logger.Info($"TestAdapter: {this.AdapterVersion}", 0);
             AdapterLogger.Logger.Info($"AdapterSettings: {AdapterSettings.Settings}");
-            //AdapterLogger.Logger.Debug($"SettingsXml: {discoveryContext.RunSettings.SettingsXml}");
+
+            if (AdapterSettings.Instance.RunSettings is null)
+            {
+                AdapterLogger.Logger.Error($"RunSettings is null");
+            }
+
+#if DEBUG
+            AdapterLogger.Logger.Warning($"TestAdapter: {this.AdapterVersion}");
+            AdapterLogger.Logger.Warning($"AdapterSettings: {AdapterSettings.Settings}");
+
+            AdapterLogger.Logger.Warning("-");
+            var collection = Metadatas.XmlUtils.ParseKeyValues(discoveryContext.RunSettings.SettingsXml);
+            foreach (var item in collection)
+            {
+                AdapterLogger.Logger.Warning($"{item.Key}: {item.Value}");
+            }
+            AdapterLogger.Logger.Warning("-");
+
+            foreach (var item in Metadatas.MapperKey.GetNames(AdapterSettings.Instance))
+            {
+                AdapterLogger.Logger.Warning($"{item}");
+            }
+            AdapterLogger.Logger.Warning("-");
+
+            //var RunSettings = Metadatas.Mapper.Map(collection, new RunSettingsModel());
+            //AdapterLogger.Logger.Warning($"RunSettings: {RunSettings}");
+            // AdapterLogger.Logger.Warning($"SettingsXml: {discoveryContext.RunSettings.SettingsXml}");
+#endif
         }
     }
 }